SPOI-6949 - Removed AbstractBucketManagerOptimized. Moved bucket collation to AbstractBucketManager. Added unit test.
